Add Windows support when using eventlet with monkey_patch
Currently when eventlet monkey patch is used and the os module
is monkey patched, os.fdopen can not be used on Windows.
This happens because Windows pipes don't support non-blocking I/O.

This patch addreses this issue by using an event for
the alert instead of a pipe. The OVS poller on Windows
supports waiting on events.

The code from connection.py was split in platform independent
and platform dependent code. There are three new folders added
which contain the code from connection.py that was splitted.
