+++ neutron/agent/ovsdb/native/vlog.py
@@ -12,6 +12,8 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
+import sys
+
 from oslo_log import log as logging
 from ovs import vlog
 
@@ -20,6 +22,17 @@ LOG = logging.getLogger(__name__)
 
 def use_oslo_logger():
     """Replace the OVS IDL logger functions with our logger"""
+
+    if sys.platform == 'win32':
+        # NOTE(abalutoiu) When using oslo logging we need to keep in mind that
+        # it does not work well with native threads. We need to be careful when
+        # we call eventlet.tpool.execute, and make sure that it will not use
+        # the oslo logging, since it might cause unexpected hangs if
+        # greenthreads are used. On Windows we have to use
+        # eventlet.tpool.execute for a call to the ovs lib which will use
+        # vlog to log messages. We will skip replacing the OVS IDL logger
+        # functions on Windows to avoid unexpected hangs with oslo logging
+        return
 
     # NOTE(twilson) Replace functions directly instead of subclassing so that
     # debug messages contain the correct function/filename/line information
